requirements.txt
pydantic>=2.6
python-dotenv>=1.0
openai>=1.35.7
rich>=13.7.1
loguru>=0.7.2
pytest>=8.2.0
numpy>=1.26
imageio>=2.34
